Update index.html
let's update this at the top to allow for browsing to and selecting a csv file with my screener results.  from there, only select for evaluation the tickers that show:

Tech Rating and MA Rating at Strong Buy
At least the last 3 trading days with greater than 5% gain
Average gain over the last 5 days of at least 8%

Remove the early portion about the 30 minute period results

Show a summary table at the top of the tickers that meet that criteria.  Allow for row selection that then leads to the data you currently show